LayoutTests/fast/text-extraction/debug-text-extraction-markdown.html

<!-- webkit-test-runner [ useFlexibleViewport=true textExtractionEnabled=true ] -->
<!DOCTYPE html>
<html>
<meta name="viewport" content="width=device-width, initial-scale=1">
<meta charset="utf-8">
<head>
<style>
body.done {
    white-space: pre-wrap;
}

.linethrough {
    text-decoration: line-through;
}
</style>
<script src="../../resources/ui-helper.js"></script>
</head>
<body>
<p>Here’s to the crazy ones. The misfits. The rebels. The troublemakers. The round pegs in the square holes.</p>
<a href="https://www.apple.com">T[e]st link</a>
<div>You can quote them, disagree with them, glorify or vilify them. About the only thing you can’t do is ignore them. Because they change things.</div>
<img src="file:///fake/image.png" alt="Purposefully broken image"></img>
<p>This is a list:</p>
<ul>
    <li>foo</li>
    <li>bar</li>
    <li>baz</li>
</ul>
On <a href="https://webkit.org">sale</a> for
<p>
    <sup>£</sup>
    10
    <sup>99</sup>
    This <strike>text</strike> has a <del>line through it</del>. The price <s>was ~~$50</s>
</p>
<p><span class="linethrough">CSS linethrough</span> normal text</p>
<p class="linethrough">Linethrough <span>also linethrough</span></p>
<p><del class="linethrough">both semantic and CSS</del></p>
<p class="linethrough"><a href="https://www.apple.com">A link</a></p>
<a href="https://www.example.com"><img src="file:///fake/linked-image.png" alt="A linked image"></a>
<script>
addEventListener("load", async () => {
    if (!window.testRunner)
        return;

    testRunner.dumpAsText();
    testRunner.waitUntilDone();

    document.body.textContent = await UIHelper.requestDebugText({
        includeURLs: true,
        outputFormat: "markdown",
    });
    document.body.classList.add("done");

    testRunner.notifyDone();
});
</script>
</body>
</html>
