LayoutTests/fast/text-extraction/debug-text-extraction-markdown-expected.txt

Here’s to the crazy ones. The misfits. The rebels. The troublemakers. The round pegs in the square holes.
[T\[e\]st link](https://www.apple.com/)
You can quote them, disagree with them, glorify or vilify them. About the only thing you can’t do is ignore them. Because they change things.
![Purposefully broken image](file:///fake/image.png)
This is a list:
- foo
- bar
- baz
On [sale](https://webkit.org/) for
£10.99 This ~~text~~ has a ~~line through it~~. The price ~~was \~\~$50~~
~~CSS linethrough~~ normal text
~~Linethrough also linethrough~~
~~both semantic and CSS~~
~~[A link](https://www.apple.com/)~~
![A linked image](file:///fake/linked-image.png) [](https://www.example.com/)
